M01A06 – Conversão de Tipos em Kotlin

+++ M01A06convertTipos/src/M01A06convertTipos.kt
@@ -1,5 +1,5 @@
 fun main(args: Array<String>) {
-    var n1:Int = 7
+    var n1:Short = 7
     var n2:Int
 
     n2 = n1
